# Fix learning rate scheduler performance issue