fixes #12260 

ResNet50 with piecewise_decay
time cost of `exe.run()`:

|               | 1 GPU |  4 GPU  |  8 GPU |
|--------|--------|---------|---------| 
|before  | 0.75s   |   1.41s    |  1.52s     |
|after     | 0.75s   |   0.88s   |  0.96s    |