@JiayiFeng 2018-07-19
恩，这块还在调，稳定后会发一个在ResNet50上的结果